+++ astro/src/layouts/Site.astro
@@ -10,7 +10,7 @@ import ContentFooter from '@layouts/content/Footer.astro';
 import type { StructuredData } from '@type/types.js';
 import { getNoteData, getReferenceData } from '@util/footnote.js';
 import { getPageUrl } from '@util/ssr.js';
-import TocUtil from '@util/Toc.js';
+import { getData as getTocData } from '@util/toc.js';
 
 interface Props {
 	astroFilePath: string | undefined;
@@ -30,7 +30,7 @@ const rendered = await Astro.slots.render('default');
 const dom = new JSDOM(rendered);
 const { document } = dom.window;
 
-const tocData = toc ? TocUtil.getData(document) : undefined;
+const tocData = toc ? getTocData(document) : undefined;
 const footnoteNoteData = getNoteData(document);
 const footnoteReferenceData = getReferenceData(document);
